# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TodoTalk messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all registered users
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: All registered users (excluding me)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
  /search:
    get:
      operationId: globalSearch
      tags: [users]
      summary: Search public users, STAR profiles, and public feed posts
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 2, maxLength: 80 }
        - in: query
          name: type
          schema: { type: string, enum: [all, users, stars, posts], default: all }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 30, default: 20 }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalSearchResponse"
        "400": { description: Invalid search query }
        "401": { description: Unauthorized }
  /search/trending:
    get:
      operationId: getTrendingSearches
      tags: [users]
      summary: Get cached popular search terms
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 10, default: 5 }
      responses:
        "200":
          description: Popular search terms
          content:
            application/json:
              schema:
                type: object
                required: [items, generatedAt]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [term, rank, change, resultCount]
                      properties:
                        term: { type: string }
                        rank: { type: integer }
                        change: { type: integer }
                        resultCount: { type: integer }
                  generatedAt: { type: string, format: date-time }
        "401": { description: Unauthorized }
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get my profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: My profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update my profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteMe
      tags: [users]
      summary: Delete my account
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Deleted
  /users/me/profile-history:
    get:
      operationId: listMyProfileHistory
      tags: [users]
      summary: List my profile update history
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Profile update history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileUpdateHistoryItem"
  /users/me/push-token:
    post:
      operationId: registerPushToken
      tags: [users]
      summary: Register push token
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenInput"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Find a user by exact email address
      security: [{ bearerAuth: [] }]
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
...
  /users/{userId}/posts:
    get:
      tags: [users]
      summary: List a user's public feed posts with cursor pagination
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: starId
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: cursor
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Public posts page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfilePostsPage" }
  /users/{userId}/growth-records:
    get:
      tags: [users]
      summary: List public growth records with cursor pagination
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
        - { name: cursor, in: query, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Growth records page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrowthRecordsPage" }
  /users/{userId}/battle-results:
    get:
      tags: [users]
      summary: List public battle results with cursor pagination
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
        - { name: cursor, in: query, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Battle results page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BattleResultsPage" }
...
  /users/{userId}/profile:
    get:
      tags: [users]
      summary: Get a public user profile
      security: [{ bearerAuth: [] }]
      parameters: [{ name: userId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Public profile, content: { application/json: { schema: { $ref: "#/components/schemas/PublicProfile" } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StarFeedAuthor:
      type: object
      required: [nickname]
      properties:
        id:
          type: ["string", "null"]
        nickname:
          type: string
        profileImageUrl:
          type: ["string", "null"]
...
    StarFeedPost:
      type: object
      required: [id, kind, title, body, visibility, createdAt, author, targetStarProfile, reactionCount, commentCount, reactedByMe, recentComments]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [official, event, fan, star, growth, profile_update, talk_diary]
        title:
          type: string
        body:
          type: string
        metadata:
          type: ["object", "null"]
          additionalProperties: true
        visibility:
          type: string
          enum: [PRIVATE, FRIENDS, PUBLIC]
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/StarFeedAuthor"
        targetStarProfile:
          oneOf:
            - $ref: "#/components/schemas/StarProfileTarget"
            - type: "null"
        reactionCount:
          type: integer
        commentCount:
          type: integer
        reactedByMe:
          type: boolean
        recentComments:
          type: array
          items:
            $ref: "#/components/schemas/StarFeedComment"
...
    UserProfile:
      type: object
      required: [id, clerkId, email, nickname, notificationEnabled, createdAt]
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        nickname:
          type: string
        profileImageUrl:
          type: ["string", "null"]
        statusMessage:
          type: ["string", "null"]
        pushToken:
          type: ["string", "null"]
        notificationEnabled:
          type: boolean
        createdAt:
          type: string
...
    UserProfileUpdate:
      type: object
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 30
        statusMessage:
          type: ["string", "null"]
          maxLength: 200
        profileImageUrl:
          type: ["string", "null"]
          maxLength: 2048
        notificationEnabled:
          type: boolean
        talkAnalysisEnabled:
          type: boolean
    ProfileUpdateHistoryItem:
      type: object
      required: [id, userId, kind, isVisible, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        kind:
          type: string
          enum: [profile_image, status_message, profile_update]
        oldProfileImageUrl:
          type: ["string", "null"]
        newProfileImageUrl:
          type: ["string", "null"]
        oldStatusMessage:
          type: ["string", "null"]
        newStatusMessage:
          type: ["string", "null"]
        feedPostId:
          type: ["string", "null"]
        isVisible:
          type: boolean
        createdAt:
          type: string
    PushTokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 8192
    GlobalSearchResponse:
      type: object
      required: [query, type, users, starProfiles, posts, nextCursor]
      properties:
        query: { type: string }
        type: { type: string, enum: [all, users, stars, posts] }
        users:
          type: array
          items:
            type: object
            required: [id, nickname, profileImageUrl, statusMessage, isMe]
            properties:
              id: { type: string, format: uuid }
              nickname: { type: string }
              profileImageUrl: { type: [string, "null"] }
              statusMessage: { type: [string, "null"] }
              isMe: { type: boolean }
        starProfiles:
          type: array
          items:
            type: object
            required: [id, displayName, starKey, imageUrl, stage, ownerId, followedByMe]
            properties:
              id: { type: string, format: uuid }
              displayName: { type: string }
              starKey: { type: string }
              imageUrl: { type: [string, "null"] }
              stage: { type: string }
              ownerId: { type: [string, "null"], format: uuid }
              followedByMe: { type: boolean }
        posts:
          type: array
          items:
            type: object
            required: [id, title, body, kind, createdAt]
            properties:
              id: { type: string, format: uuid }
              title: { type: string }
              body: { type: string }
              kind: { type: string }
              createdAt: { type: string, format: date-time }
        nextCursor: { type: [string, "null"] }
    PublicUser:
      type: object
      required: [id, nickname, accountKind]
      properties:
        id:
          type: string
        nickname:
          type: string
        accountKind:
          type: string
          enum: [user, official, system]
        friendAlias:
          type: ["string", "null"]
          description: Current viewer's saved name for this user, when available.
        displayName:
          type: string
          description: friendAlias when set, otherwise nickname.
        profileImageUrl:
          type: ["string", "null"]
        statusMessage:
          type: ["string", "null"]
...
    ProfilePostsPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/StarFeedPost" }
        nextCursor:
          type: string
          format: date-time
          nullable: true
    GrowthRecordsPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/GrowthRecord" }
        nextCursor: { type: string, format: date-time, nullable: true }
...
    BattleResultsPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PublicBattleResult" }
        nextCursor: { type: string, format: date-time, nullable: true }
...
    PublicProfile:
      type: object
      required: [id, nickname, stars, followerCount, followingCount]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string }
        profileImageUrl: { type: string, nullable: true }
        statusMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        fan: { type: object }
        stars: { type: array, items: { type: object } }
        followerCount: { type: integer }
        followingCount: { type: integer }
        followedStarIds: { type: array, items: { type: string, format: uuid } }